# Use faster serializers for Search::Postgres::Username