--- a/app/serializers/search/username_serializer.rb
+++ b/app/serializers/search/username_serializer.rb
@@ -1,0 +1,8 @@
+module Search
+  class UsernameSerializer < ApplicationSerializer
+    attributes :id,
+               :name,
+               :profile_image_90,
+               :username
+  end
+end