<!--
     For Work In Progress Pull Requests, please use the Draft PR feature,
     see https://github.blog/2019-02-14-introducing-draft-pull-requests/ for further details.

     For a timely review/response, please avoid force-pushing additional
     commits if your PR already received reviews or comments.

     Before submitting a Pull Request, please ensure you've done the following:
     - 📖 Read the Forem Contributing Guide: https://github.com/forem/forem/blob/master/CONTRIBUTING.md#create-a-pull-request.
     - 📖 Read the Forem Code of Conduct: https://github.com/forem/forem/blob/master/CODE_OF_CONDUCT.md.
     - 👷‍♀️ Create small PRs. In most cases this will be possible.
     - ✅ Provide tests for your changes.
     - 📝 Use descriptive commit messages.
     - 📗 Update any related documentation and include any relevant screenshots.

     NOTE: Pull Requests from forked repositories will need to be reviewed by
     a Forem Team member before any CI builds will run. Once your PR is approved
     with a `/ci` reply to the PR, it will be allowed to run subsequent builds without
     manual approval.
-->

## What type of PR is this? (check all applicable)
- [x] Optimization

## Description
Inspired by the findings of @rhymes in https://github.com/forem/forem/pull/13022, this PR updates `Search::Postgres::Username` to use a serializer over `.as_json`.

```ruby
require 'benchmark/ips'

data = User.search_by_username("a").limit(6).select(%i[id name profile_image username])

Benchmark.ips do |x|
  x.report("as_json") do
    data.as_json
  end
  
  x.report("jsonapi-serializer") do
    Search::UsernameSerializer.new(data, is_collection: true).serializable_hash[:data].pluck(:attributes)
  end
  
  x.compare!
end
```

```
Warming up --------------------------------------
             as_json   237.000  i/100ms
  jsonapi-serializer   828.000  i/100ms
Calculating -------------------------------------
             as_json      2.406k (± 4.7%) i/s -     12.087k in   5.035621s
  jsonapi-serializer      7.780k (± 4.2%) i/s -     38.916k in   5.010714s

Comparison:
  jsonapi-serializer:     7780.5 i/s
             as_json:     2405.6 i/s - 3.23x  (± 0.00) slower
```

This more than 3x faster 🎉 .

## Related Tickets & Documents
https://github.com/forem/forem/pull/12975

## QA Instructions, Screenshots, Recordings
N/A - specs will do!

### UI accessibility concerns?
N/A

## Added tests?
- [x] Yes

## [Forem core team only] How will this change be communicated?
- [x] I will share this change internally with the appropriate teams

![must_go_faster_jurassic_park_gif](https://media.giphy.com/media/7XsFGzfP6WmC4/giphy.gif)